requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-socketio==5.11.3
eventlet==0.36.1
python-dotenv==1.0.1
Flask-Cors==4.0.1
PyJWT==2.9.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
Flask-Migrate==4.0.7


